requirements.txt
streamlit==0.82.0
plotly==4.14.3
pandas==1.1.5
numpy==1.19.4
Pillow==8.2.0
